Added Test Target
* Took Void Safety back to Conformance, where everything is compiling.
* Added test target and modified ECF to take advantage of it.

NOTE: The primary el_base target will not have the "tests" folder (cluster), but it has a new library called "test_extension" (get this off of GitHib--search for "test_extension" under ljr1981 user ID).

The "test_extension" library ought to be cloned to the same root as Eiffel-loop so that the indirect reference (e.g. ..\,,\,,\) will work properly.

The "test_extension" library allows for feature level exports of features designed in the class to be seen by the testing target only. Otherwise, you DO NOT want production code having the ability to see, call, or access test-only related features even though they appear in library code. In other words--we never want to ship test-only code as though it is production, which would allow reuse-consumers to get dependency on that code, and then if we change or remove the testing, then suddenly it breaks client systems! That would be a no-no!